update: lots of additional comments

+++ mimicle/Assets/Scripts/Wave/Waves/WaveData.cs
@@ -7,6 +7,9 @@ namespace Self.Game
 
     public class WaveData : MonoBehaviour
     {
+        /// <summary>
+        /// 開始ウェーブ
+        /// </summary>
         public Activate waves = Activate.First;
 
         [SerializeField]
@@ -16,15 +19,30 @@ namespace Self.Game
         WaveUI waveUI;
 
         int max = 0;
+        /// <summary>
+        ///  最大ウェーブ数
+        /// </summary>
         public int Max => max;
 
         int current4ui = 0;
-
+        /// <summary>
+        /// アクティブウェーブ(UI表示用)
+        /// </summary>
         public int Current4UI => current4ui;
+
+        /// <summary>
+        /// アクティブウェーブ
+        /// </summary>
         public int CurrentActive => current4ui - 1;
 
+        /// <summary>
+        /// ゲームクリアしたらtrue
+        /// </summary>
         public bool IsDone { get; set; }
 
+        /// <summary>
+        /// アクティブウェーブとwaveが同じならtrue
+        /// </summary>
         public bool IsActiveWave(int wave) => wave == CurrentActive;
 
         void Start()
@@ -36,12 +54,15 @@ namespace Self.Game
             waveUI.Start();
         }
 
+        /// <summary>
+        /// 次のウェーブに進む
+        /// </summary>
         public void Next()
         {
             current4ui = ((int)waves) + 1;
             waveObjs[current4ui - 1].SetActive(true);
 
-            for (ushort index = 0; index < waveObjs.Length; index++)
+            for (int index = 0; index < waveObjs.Length; index++)
             {
                 if (current4ui - 1 != index)
                 {
@@ -50,6 +71,10 @@ namespace Self.Game
             }
         }
 
+        /// <summary>
+        /// ウェーブnをアクティブにする
+        /// </summary>
+        /// <param name="n"></param>
         public void ActivateWave(int n)
         {
             waveObjs[n].SetActive(true);
